+++ CHANGELOG.md
@@ -17,6 +17,9 @@ The backward compatibility promise has the following exceptions:
 
 ## [3.0.0] - 2023-08-22
 ### Added
+- Reply tube selections when reconnecting a failed Pool connection.
+
+### Added
 - Add new `BuriedException` thrown by `put()` and `release()` when the server
   returns this error. This is a minor **BC break** because these commands
   previously returned a positive response for this error.